IRMATube movies now have real access restrictions
The implementation reimplements the default webserver, and makes sure that movies can only be streamed when the user is properly authenticated via IRMA. Please see data/README for the details.

Also, irma_configuration should henceforth be linked into data, to keep the directory structure cleaner. The eclipse settings reflect this.